feature: Realizado todas funcionalidade encontradas como editadas no arquivo requisitos.md

+++ Recaraga.java
@@ -2,17 +2,17 @@
 import java.time.Instant;
 
 public class Recaraga {
-    private final float kWh;
+    private final double kWh;
     private final Instant data;
     private final Posto posto;
 
-    public Recaraga(float kWh, Instant data, Posto posto) {
+    public Recaraga(double kWh, Instant data, Posto posto) {
         this.kWh = kWh;
         this.data = data;
         this.posto = posto;
     }
 
-    public float getkWh() {
+    public double getkWh() {
         return kWh;
     }
